added a file input to Profile Update button and cleanup UserSettingUpdateModel file
