+++ frontend/src/data/articles.js
@@ -7,10 +7,7 @@ const articles = [
     body:
       "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
     slug: 'article1',
-    tags:[
-      "Python",
-      "JavaScript"
-    ]
+    tags: ['Python', 'JavaScript'],
   },
   {
     id: '2',
@@ -20,11 +17,7 @@ const articles = [
     body:
       "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
     slug: 'article2',
-    tags:[
-      "Python",
-      "C#",
-      'Google Maps'
-    ]
+    tags: ['Python', 'C#', 'Google Maps'],
   },
   {
     id: '3',
@@ -34,10 +27,7 @@ const articles = [
     body:
       "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
     slug: 'article3',
-    tags:[
-      'React JS',
-      'CSS3'
-    ],
+    tags: ['React JS', 'CSS3'],
   },
 ];
